scripts/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
umap-learn>=0.5
shap>=0.43
requests>=2.31
networkx>=3.1
ruptures>=1.1
